Committing cleanup
Reduced number of test files to just: test_chk01_checklist,
test_epsig_log_files, test_file_name_format, test_msl_file_content,
test_psl_file_content, test_tsl_file_content.

Removed: test_file_format, test_general_file_format and test_PSL_files.

Duplicate unit tests were occurring grouped up similar tests.
