+++ README.md
@@ -30,7 +30,7 @@ Main class (QCASTestClient) derived from unittest.TestCase, includes the followi
 
 The QCASTestClient includes all modules that have common test procedures. 
 
-## Module: `test_chk01_checklist.py`
+## CHK01: Datafiles Checklist Module: `test_chk01_checklist.py`
 Mirrors the checks specified in CHK01 Casino Datafiles checklist, the following unit tests are performed in this test: 
 
 #### `test_Generated_PSL_files_Differ()`
@@ -59,8 +59,14 @@ Verifies the following entry for each new game generated:
 - Verifies that the generated PSL entries for two months is created for the Random game. 
 - Verifies that the PSL entries matches for two months exists in `self.PSLfile and self.nextMonth_PSLfile`
 
-## Module: `test_epsig_log_files.py`
+
+
+
+## Epsig Log file Verification Module: `test_epsig_log_files.py`
 This test script verifies the expected output of the EPSIG log. 
+
+#### `test_Read_Epsig_log_file_from_disk()`
+- Verifies self.my_preferences.epsig_log_file can be read from disk
 
 #### `test_Epsig_Log_file()`
 - Verifies the last entry of the Epsig log file
@@ -70,7 +76,9 @@ This test script verifies the expected output of the EPSIG log.
 - Verifies that the end of the Epsig Log File indicates: "with EXIT_SUCCESS"
 - Verifies that the command that was used for Epsig is correct. (Correct Epsig Binary used; Correct BINIMAGE Path used: i.e. G:\; Correct Datafiles referenced, i.e. MSL file is `self.MSLfile or self.nextMonth_MSLfile`; TSL file is `self.TSLfile`; PSL file is `self.PSLfile or self.nextMonth_PSLfile`
 
-## Module: `test_file_name_format.py`
+
+
+## Filename Test Module: `test_file_name_format.py`
 Generic test scripts for correct file name format and conventions. 
 
 #### `test_MSL_filename_ends_with_MSL()`
@@ -79,7 +87,7 @@ Generic test scripts for correct file name format and conventions.
 #### `test_MSL_filename_date()`
 - Verifies that the month fields in `self.MSLfile or self.nextMonth_MSLfile` are not equal
 - Verifies that the MSL filename month fields are not equal
-3- Verifies that the MSL year fields is the same if month is less than 12, otherwise an increment in Year value is expected
+- Verifies that the MSL year fields is the same if month is less than 12, otherwise an increment in Year value is expected
 
 #### `test_MSL_filename_version()`
 - Verifies that the MSL version fields must always be `v1`
@@ -91,55 +99,62 @@ Generic test scripts for correct file name format and conventions.
 - Verifies that the `self.PSLfile or self.nextMonth_PSLfile` ends with '.psl'
 
 #### `test_PSL_file_version_increment()`
-- [to be completed] Verifies that the PSL file version should be incremented if not a new moth.
+- Verifies that the PSL file version should be incremented if not a new moth.
 
 #### `test_PSL_filename_date()`
 - Verify that current PSL month is not equal to the new PSL month 
 - Verify that the PSL year is the same, unless current PSL month is December.
 
-## Module: `test_general_file_format.py`
-#### `test_Read_PSL_file()`
-- Verifies that `self.PSLfile` and `self.nextMonth_PSLfile` can be read from disk
 
-#### `test_Read_TSL_file()`
+## TSL file Module: `test_tsl_file_content.py`
+
+####`test_Read_TSL_file_from_disk()`
 - Verifies that `self.TSLfile` and `self.previous_TSLfile` can be read from disk
 
-#### `test_MSL_fields()`
-- Verifies `self.MSLfile` file format
-- Verifies MSL file month field is this month's or next
-- Verifies MSL file year field is this year's or next
-- Verifies MSL file has 31 seeds
+#### `test_TSL_content_can_be_parsed()`
+- Verifies TSL file manufacturer field entries is valid
+- Verifies TSL file SSAN field is unique
+- Verifies TSL file Bin Image Type is valid 
+- Verifies BNK file name is uniqe (Note, this fails and is disabled by default as BNK file: 0101230 duplicates exist. 
 
-#### `test_PSL_fields()`
-- Verifies `self.PSLfile` and `self.nextMonth_PSLfile` file format
+## PSL file Module: `test_psl_file_content.py`
+
+#### `test_psl_size_is_reasonable()`
+- Verifies that the file size of the PSL files is reasonable (greater than 1055KB as at July 2013)
+
+### `test_PSL_content_can_be_parsed()`
+- Verifies the `self.PSLfile` and `self.nextMonth_PSLfile` file formats
 - Verifies PSL file manufacturer field is valid
 - Verifies PSL Game name field length is 30 characters or less
 - Verifies PSL year field is this year's or next
 - Verifies PSL month field is this month's or next
-
-#### `test_TSL_fields()`
-- Verifies TSL file manufacturer field entries is valid
-- Verifies TSL file SSAN field is unique
-- Verifies TSL file Bin Image Type is valid 
-
-## Module: `test_PSL_files.py`
-#### `test_PSL_size_is_reasonable()`
-- Verifies that the file size of the PSL files is reasonable (greater than 1055KB as at July 2013)
-
-### `test_PSL_content()`
-- Verifies the `self.PSLfile` and `self.nextMonth_PSLfile` file formats
 - Verifies that the PSL file has the same number of lines as the generated PSL entry
 
-## Module: `test_MSL_files.py`
+#### `test_Read_PSL_file_from_disk()`
+- Verifies that `self.PSLfile` and `self.previous_PSLfile` can be read from disk
+
+#### `test_valid_MIDs_have_PSL_entries()`
+- Verifies that each Manufacturer has a PSL entry, on each PSL file
+
+#### `test_date_field_in_PSL_entry_equals_date_field_in_filename()`
+- Verifies that the date in the first PSL entry has the correct month & year, for both PSL files. 
+
+## MSL file Module: `test_msl_file_content.py`
 #### `test_MSL_size_is_reasonable()`
 - Verifies that the file size of the MSL files is reasonable (The size should not change and is 1KB)
 
-#### `test_MSL_content()`
-- Verifies the `self.MSLfile` and `self.nextMonth_MSLfile` file formats
-- Verifies that the MSL files only has one entry
+#### `test_MSL_content_can_be_parsed()`
+- Verifies the `self.MSLfile` and `self.nextMonth_MSLfile` file contents by reading parsing the files
+- Verifies that the MSL object generated from the files only has one entry
 
 #### `test_MSL_file_one_row()`
 - Verifies that the MSL files only has one entry
 
 #### `test_Read_MSL_file()`
 - Verifies that the `self.MSLfile` and `self.nextMonth_MSLfile` can be read from disk 
+
+#### `test_MSL_fields_sanity_checks()`
+- Performs sanity checks on MSL fields, for both MSL files
+- Check Month Field to be either this month's or next month's, for both MSL files
+- Check Year Field to be either this month's or next month's, for both MSL files
+- Check number of Seeds equal 31, for both MSL files
